Fix for bug 930.
Calculate with doubles rather than float.
Simplify calculation of max/min peak.
Slight hack to work around limitation in FloatingPointValidator.
Allow 'close equality' in CheckClip().
Tested on Debian 8 and Release build on Ubuntu 14.10.

+++ src/effects/Amplify.h
@@ -68,11 +68,11 @@ private:
    void CheckClip();
 
 private:
-   float mPeak;
+   double mPeak;
 
-   float mRatio;
-   float mAmp;
-   float mNewPeak;
+   double mRatio;
+   double mAmp;
+   double mNewPeak;
    bool mCanClip;
 
    wxSlider *mAmpS;
